requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.112.0
h11==0.14.0
idna==3.7
numpy==2.0.1
openpyxl==3.1.5
panda==0.3.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.3
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
